Split webpack config file to two files (simplicity)
Passing parameters to webpack configuration is redunant. The easier
solution is split the config into two of them. In the future, I can use
that approach of passing paramter by --target or corss-env for passing
build type (debug, release). For now, I think it is better to keep it
simple.